Moved bottom components to their own pages, migrated universal data to singleton
