+++ ride_on/ride_on/lib/objects/rideObject.dart
@@ -34,17 +34,15 @@ class RideObject implements Comparable<RideObject>
     //vehicle
   DateTime rideDate; //ride date
 
-  void addPoint(LatLng newPoint) { rideRoute.add(newPoint); }
-  void setMax(double newMax) { maxSpeed = (mpsTomph(newMax) > maxSpeed) ? mpsTomph(newMax) : maxSpeed; }
-  void addDistance(double distanceIn) { rideLength += distanceIn; }
-  double getDistance() { return rideLength * .000621371; }
-  void incRideTime() { rideTimeSec++; }
-  double getAvgSpeed() { return mpsTomph(rideLength / rideTimeSec);}
-  void setDate(DateTime dateIn) {
-    rideDate = dateIn; }
-  void setName(String name) { vehicleName = name; }
-  void setUserID(String userID) {
-    this.userId = userID;}
+  void addPoint(LatLng newPoint) {rideRoute.add(newPoint);}
+  void setMax(double newMax) {maxSpeed = (mpsTomph(newMax) > maxSpeed) ? mpsTomph(newMax) : maxSpeed;}
+  void addDistance(double distanceIn) {rideLength += distanceIn;}
+  double getDistance() {return rideLength * .000621371;}
+  void incRideTime() {rideTimeSec++;}
+  double getAvgSpeed() {return mpsTomph(rideLength / rideTimeSec);}
+  void setDate(DateTime dateIn) {rideDate = dateIn;}
+  void setName(String name) {vehicleName = name;}
+  void setUserID(String userID) {this.userId = userID;}
   void setRideLength(double length) {this.rideLength = length;}
   void setRideTime(int time) {this.rideTimeSec = time;}
   void setRideRouteDoubles (List<double> route)
@@ -122,7 +120,6 @@ class RideObject implements Comparable<RideObject>
         else
         {
           lat = coordinate;
-
         }
         i++;
       }
@@ -136,5 +133,4 @@ class RideObject implements Comparable<RideObject>
     }
     else return 1;
   }
-
 }
